+++ dd-java-agent/instrumentation/spring-web/src/main/java/datadog/trace/instrumentation/springweb/SpringWebErrorInstrumentation.java
@@ -1,0 +1,62 @@
+package datadog.trace.instrumentation.springweb;
+
+import static io.opentracing.log.Fields.ERROR_OBJECT;
+import static net.bytebuddy.matcher.ElementMatchers.isInterface;
+import static net.bytebuddy.matcher.ElementMatchers.isMethod;
+import static net.bytebuddy.matcher.ElementMatchers.isProtected;
+import static net.bytebuddy.matcher.ElementMatchers.nameStartsWith;
+import static net.bytebuddy.matcher.ElementMatchers.named;
+import static net.bytebuddy.matcher.ElementMatchers.not;
+import static net.bytebuddy.matcher.ElementMatchers.takesArgument;
+
+import com.google.auto.service.AutoService;
+import datadog.trace.agent.tooling.Instrumenter;
+import io.opentracing.Scope;
+import io.opentracing.Span;
+import io.opentracing.tag.Tags;
+import io.opentracing.util.GlobalTracer;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.Map;
+import net.bytebuddy.asm.Advice;
+import net.bytebuddy.description.type.TypeDescription;
+import net.bytebuddy.matcher.ElementMatcher;
+
+@AutoService(Instrumenter.class)
+public final class SpringWebErrorInstrumentation extends Instrumenter.Default {
+
+  public SpringWebErrorInstrumentation() {
+    super("spring-web");
+  }
+
+  @Override
+  public ElementMatcher<TypeDescription> typeMatcher() {
+    return not(isInterface()).and(named("org.springframework.web.servlet.DispatcherServlet"));
+  }
+
+  @Override
+  public Map<ElementMatcher, String> transformers() {
+    final Map<ElementMatcher, String> transformers = new HashMap<>();
+    transformers.put(
+        isMethod()
+            .and(isProtected())
+            .and(nameStartsWith("processHandlerException"))
+            .and(takesArgument(3, Exception.class)),
+        SpringWebErrorHandlerAdvice.class.getName());
+    return transformers;
+  }
+
+  public static class SpringWebErrorHandlerAdvice {
+    @Advice.OnMethodEnter(suppress = Throwable.class)
+    public static void nameResource(@Advice.Argument(3) final Exception exception) {
+      final Scope scope = GlobalTracer.get().scopeManager().active();
+      if (scope != null && exception != null) {
+        final Span span = scope.span();
+        span.log(Collections.singletonMap(ERROR_OBJECT, exception));
+        // We want to capture the stacktrace, but that doesn't mean it should be an error.
+        // We rely on a decorator to set the error state based on response code. (5xx -> error)
+        Tags.ERROR.set(span, false);
+      }
+    }
+  }
+}
